requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.2.1
aioredis==2.0.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.8.0
anytree==2.12.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
bce-python-sdk==0.9.29
beautifulsoup4==4.12.3
cachetools==5.5.2
camel-converter==4.0.1
certifi==2025.1.31
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dashscope==1.14.1
debugpy==1.8.12
decorator==5.2.1
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
Django==5.1.6
dnspython==2.7.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.7.4
fastjsonschema==2.21.1
filelock==3.17.0
fire==0.4.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gitdb==4.0.12
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-auth==2.38.0
google-generativeai==0.4.1
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.1
hyperframe==6.1.0
idna==3.10
imap-tools==1.5.0
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
javalang==0.13.0
jedi==0.19.2
jieba==0.42.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
lancedb==0.4.0
lazy-object-proxy==1.10.0
libcst==1.0.1
loguru==0.6.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
meilisearch==0.21.0
metagpt==0.8.1
more-itertools==10.6.0
motor==3.7.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.24.3
openai==1.6.1
openapi-core==0.18.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.1.1
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.50.0
portalocker==2.10.1
prance==23.6.21.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.6.4
pydantic_core==2.16.3
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.8.0
pylance==0.9.0
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.2.1
qdrant-client==1.7.0
qianfan==0.3.2
rank-bm25==0.2.2
ratelimiter==1.2.0.post0
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.3.2
scipy==1.15.2
semantic-kernel==0.4.3.dev0
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
ta==0.10.2
tenacity==8.2.3
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.15.0
yarl==1.18.3
zhipuai==2.0.1
